+++ test/Test/Iris.hs
@@ -6,6 +6,7 @@ import Test.Iris.Cli (cliSpec, cliSpecParserConflicts)
 import Test.Iris.Colour (colourSpec)
 import Test.Iris.Interactive (interactiveSpec)
 import Test.Iris.Tool (toolSpec)
+import Test.Iris.IO (ioSpec)
 
 irisSpec :: Spec
 irisSpec = describe "Iris" $ do
@@ -14,3 +15,4 @@ irisSpec = describe "Iris" $ do
     colourSpec
     toolSpec
     interactiveSpec
+    ioSpec